+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/ArccValidationResponse.java
@@ -1,6 +1,7 @@
 package org.grnet.cat.dtos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotNull;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
@@ -21,4 +22,24 @@ public class ArccValidationResponse {
         )
         @JsonProperty("is_valid")
         public boolean isValid;
+
+        @Schema(
+                type = SchemaType.STRING,
+                implementation = String.class,
+                description = "Date and Time when the automated test ran",
+                example = " 2023-06-09 12:19:31.333059"
+        )
+        @JsonProperty("last_run")
+        public String lastRun;
+
+        @Schema(
+                type = SchemaType.STRING,
+                implementation = String.class,
+                description = "The reason of the failed check",
+                example = " 2023-06-09 12:19:31.333059"
+        )
+        @JsonProperty("message")
+        @NotNull
+        public String message;
+
 }